+++ pkg/providers/dns/testdata/TestResDnsZone/create_multisigner.tf
@@ -1,0 +1,16 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+resource "akamai_dns_zone" "multi_signer_test_zone" {
+  contract       = "ctr1"
+  zone           = "multisignerexampleterraform.io"
+  type           = "primary"
+  comment        = "This is a test zone with multi-signer DNSSEC"
+  sign_and_serve = false
+  group          = "grp1"
+
+  multi_provider_dnssec {
+    enabled = true
+  }
+}